Maintenance: Code refactoring
