+++ backend/api/cert/urls.py
@@ -2,13 +2,13 @@ from django.urls import path, include
 from rest_framework.routers import DefaultRouter
 from api.cert.views.CertificateView import CertificateViewSet
 from api.cert.views.searchView import search
-from api.cert.views.IssueView import issue
+from api.cert.views.createView import create
 
 router = DefaultRouter()
 router.register('', CertificateViewSet, basename='cert')
 
 urlpatterns = [
-    path('issue/', issue, name='issue'),
+    path('create/', create, name='create'),
     path('search/', search, name='search'),
     path('', include(router.urls)), 
 ]
