+++ backend/api/edu/serializers/__init__.py
@@ -1,5 +1,7 @@
 from .EducationCenterSerializer import EducationCenterSerializer
+from .EducationCenterSummarySerializer import EducationCenterSummarySerializer
 
 __all__ = [
     "EducationCenterSerializer",
+    "EducationCenterSummarySerializer"
 ]
