+++ frontend/src/features/certificate/services/cert.api.ts
@@ -1,22 +1,33 @@
 import api from '../../../libs/axios';
 import { Certificate } from '@/features/certificate/types/Certificate.type';
 
-export const fetchCertificates = (params: any) => 
-  api.get('/cert/', { params }).then(res => res.data);
-
-export const fetchCertificateByUuid = (uuid: string) =>
-  api.get(`/cert/${uuid}/`).then(res => res.data);
-
-export const createCertificate = (data: FormData) => {
-  return api.post('/cert/', data, {
-    headers: { 'Content-Type': 'multipart/form-data' },
-  });
-};
-
-export const updateCertificate = async (uuid: string, data: Partial<Certificate>): Promise<Certificate> => {
-  const response = await api.patch<Certificate>(`/cert/${uuid}/`, data);
+// 전체 자격증 조회 API
+export const fetchAllCertificate = async (): Promise<Certificate[]> => {
+  const response = await api.get('/cert/');
   return response.data;
 };
 
-export const issueCertificates = (uuids: string[]) =>
-  api.post('/cert/issue/', { uuids }).then(res => res.data);
+// 단일 자격증 상세조회 API
+export const fetchCertificateByUuid = async (uuid: string): Promise<Certificate[]> => {
+  const response = await api.get(`/cert/${uuid}/`).then(res => res.data);
+  return response.data;
+}
+
+// 자격증 조건검색 API
+export const searchCertificates = async (params: { filter_type?: string; search_value?: string; edu_name?: string; session?: string; }): Promise<Certificate[]> => {
+  const response = await api.get('/cert/search/', {
+    params,
+  });
+  return response.data;
+};
+
+// 자격증 생성 API
+export const createCertificates = (uuids: string[]) => {
+  api.post('/cert/create/', { uuids }).then(res => res.data)
+}
+
+// 자격증 수정 API
+export const updateCertificates = async (uuid: string, data: Partial<Certificate>): Promise<Certificate> => {
+  const response = await api.patch<Certificate>(`/cert/${uuid}/`, data);
+  return response.data;
+};
