+++ frontend/src/features/certificate/types/CertificateSummary.type.tsx
@@ -1,0 +1,10 @@
+export interface CertificateSummary{
+  uuid: string;         // 고유식별번호
+  course_name: string;  // 자격과정
+  issue_number: string; // 발급번호
+  issue_date: string;   // 발급일자
+  education_center?: {  // 교육원 정보
+    edu_name: string;
+    session: string;
+  };
+}
